Added a commend in the readme about downloading the file as an applescript and then saving it as an application
The script can be edited on your computer as an applescript file or as
an application file but it's easier to view on github as applescript.
If you choose to download the applescript file you'll need to open it
up in Applescript and save it as an application file (hold down the alt
key when hovering over Save). Applescript files will not launch
automatically when attached to a calander event. Only Application files
will).